+++ common/src/main/java/com/github/teamfossilsarcheology/fossil/recipe/CultureVatRecipe.java
@@ -1,5 +1,6 @@
 package com.github.teamfossilsarcheology.fossil.recipe;
 
+import com.github.teamfossilsarcheology.fossil.inventory.CultureVatMenu;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Ingredient;
@@ -8,8 +9,8 @@ import net.minecraft.world.item.crafting.RecipeType;
 import org.jetbrains.annotations.NotNull;
 
 public class CultureVatRecipe extends WithFuelRecipe {
-    protected CultureVatRecipe(ResourceLocation location, Ingredient input, Ingredient fuel, ItemStack output, int duration) {
-        super(location, input, fuel, output, duration);
+    protected CultureVatRecipe(ResourceLocation location, Ingredient input, Ingredient fuel, ItemStack output, int duration, int fuelDuration) {
+        super(location, input, fuel, output, duration, fuelDuration);
     }
 
     @Override
@@ -34,6 +35,16 @@ public class CultureVatRecipe extends WithFuelRecipe {
 
         public Serializer(Constructor<CultureVatRecipe> constructor) {
             super(constructor);
+        }
+
+        @Override
+        int defaultDuration() {
+            return CultureVatMenu.CULTIVATION_DURATION;
+        }
+
+        @Override
+        int defaultFuelDuration() {
+            return CultureVatMenu.DEFAULT_FUEL_DURATION;
         }
     }
 }