+++ forge/src/main/java/com/github/teamfossilsarcheology/fossil/forge/compat/farmers/FarmersDelightCompat.java
@@ -35,10 +35,7 @@ public class FarmersDelightCompat {
 //        manager.addPlant(ModItems.PUMPKIN_SOUP.get());
 //        manager.addPlant(ModItems.MUSHROOM_RICE.get());
 //        manager.addPlant(ModItems.VEGETABLE_NOODLES.get());
-//        manager.addPlant(ModItems.VEGETABLE_NOODLES.get());
 //        manager.addPlant(ModItems.STUFFED_PUMPKIN.get());
-//        manager.addPlant(ModItems.ROAST_CHICKEN.get());
-//        manager.addPlant(ModItems.ROAST_CHICKEN.get());
 //        manager.addPlant(ModItems.CABBAGE_SEEDS.get(), 5);
 //        manager.addPlant(ModItems.TOMATO_SEEDS.get(), 5);
 //